# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/stores": {
      "get": {
        "operationId": "StoreController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "상점 이름 검색",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "카테고리 ID로 필터링",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "description": "점주 ID로 필터링",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "활성 상태로 필터링",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isVerified",
            "required": false,
            "in": "query",
            "description": "인증 상태로 필터링",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "정렬 기준",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "name",
                "categoryId"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "정렬 방향",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상점 목록 반환"
          }
        },
        "summary": "상점 목록 조회",
        "tags": [
          "상점"
        ]
      },
      "post": {
        "operationId": "StoreController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "상점 등록 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "409": {
            "description": "사업자 등록 번호 중복"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "상점 등록",
        "tags": [
          "상점"
        ]
      }
    },
...
    "/api/stores/{storeId}/special-days": {
      "get": {
        "operationId": "SpecialDaysController_findAll",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "description": "매장 ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "조회 시작일 (YYYY-MM-DD 형식)",
            "schema": {
              "example": "2025-04-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "조회 종료일 (YYYY-MM-DD 형식)",
            "schema": {
              "example": "2025-04-30",
              "type": "string"
            }
          },
          {
            "name": "isClosed",
            "required": false,
            "in": "query",
            "description": "휴무일만 조회 (true: 휴무일만, false: 특별 영업일만, 미지정: 모두)",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특별 영업일/휴무일 목록 반환"
          },
          "404": {
            "description": "매장을 찾을 수 없음"
          }
        },
        "summary": "매장 특별 영업일/휴무일 목록 조회",
        "tags": [
          "매장 특별 영업일/휴무일"
        ]
      },
      "post": {
        "operationId": "SpecialDaysController_create",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpecialDayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "특별 영업일/휴무일 등록 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "매장을 찾을 수 없음"
          },
          "409": {
            "description": "이미 존재하는 날짜의 특별 영업일/휴무일"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 특별 영업일/휴무일 등록",
        "tags": [
          "매장 특별 영업일/휴무일"
        ]
      },
      "delete": {
        "operationId": "SpecialDaysController_removeAll",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특별 영업일/휴무일 모두 삭제 성공"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "매장을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 특별 영업일/휴무일 모두 삭제",
        "tags": [
          "매장 특별 영업일/휴무일"
        ]
      }
    },
    "/api/stores/{storeId}/special-days/{id}": {
      "get": {
        "operationId": "SpecialDaysController_findOne",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "description": "매장 ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "특별 영업일/휴무일 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특별 영업일/휴무일 정보 반환"
          },
          "404": {
            "description": "특별 영업일/휴무일을 찾을 수 없음"
          }
        },
        "summary": "매장 특별 영업일/휴무일 상세 조회",
        "tags": [
          "매장 특별 영업일/휴무일"
        ]
      },
      "patch": {
        "operationId": "SpecialDaysController_update",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpecialDayDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "특별 영업일/휴무일 수정 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "특별 영업일/휴무일을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 특별 영업일/휴무일 수정",
        "tags": [
          "매장 특별 영업일/휴무일"
        ]
      },
      "delete": {
        "operationId": "SpecialDaysController_remove",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특별 영업일/휴무일 삭제 성공"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "특별 영업일/휴무일을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 특별 영업일/휴무일 삭제",
        "tags": [
          "매장 특별 영업일/휴무일"
        ]
      }
    },
    "/api/stores/{storeId}/special-days/range": {
      "delete": {
        "operationId": "SpecialDaysController_removeByDateRange",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특별 영업일/휴무일 범위 삭제 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "매장을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 특별 영업일/휴무일 범위 삭제",
        "tags": [
          "매장 특별 영업일/휴무일"
        ]
      }
    },
...
  },
  "info": {
    "title": "픽업 서버 API",
    "description": "픽업 서버 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSpecialDayDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2025-04-10",
            "description": "특별 영업일/휴무일 날짜 (YYYY-MM-DD 형식)"
          },
          "isClosed": {
            "type": "boolean",
            "example": true,
            "description": "휴무일 여부 (true: 휴무, false: 특별 영업)"
          },
          "openingTime": {
            "type": "string",
            "example": "09:00",
            "description": "특별 영업일 오픈 시간 (휴무일이 아닌 경우 필수)"
          },
          "closingTime": {
            "type": "string",
            "example": "18:00",
            "description": "특별 영업일 마감 시간 (휴무일이 아닌 경우 필수)"
          },
          "reason": {
            "type": "string",
            "example": "창립기념일",
            "description": "특별 영업일/휴무일 사유"
          }
        },
        "required": [
          "date",
          "isClosed"
        ]
      },
      "UpdateSpecialDayDto": {
        "type": "object",
        "properties": {
          "isClosed": {
            "type": "boolean",
            "example": true,
            "description": "휴무일 여부 (true: 휴무, false: 특별 영업)"
          },
          "openingTime": {
            "type": "string",
            "example": "09:00",
            "description": "특별 영업일 오픈 시간"
          },
          "closingTime": {
            "type": "string",
            "example": "18:00",
            "description": "특별 영업일 마감 시간"
          },
          "reason": {
            "type": "string",
            "example": "창립기념일",
            "description": "특별 영업일/휴무일 사유"
          }
        }
      },
...
    }
  }
}